Abandon PostgresSessionWrapper & QueryWrapper
The original PostgresSessionWrapper (4fc5315)
was a good idea, but it's not enough to solve the problem.
Therefore QueryWrapper was added (7ccfdb2). It seems
to fix the problems, but is overly complicated and
fragile (needs to distinguish between methods
which access db and which don't).

This commit removes these wrappers and explicitly
wraps (try-except) parts of code, which access db.

Also make sure we use first() and one() correctly:
first(): returns None if no result
one(): raises NoResultFound if no result